curthooks: rework the dpkg-reconfigure
After thinking more, preseeding a database for already installed packages
and then calling 'dpkg-reconfigure' is not going to work.  debconf is
not a registry, and correct path would be for filesystem data to be
preferred.  So, for some packages, we know how to wipe configuration.

For others, we just don't reconfigure.

bzr-revno: 84.1.10