helpers: improve 'partition' to support GPT via sgdisk, fix sfdisk create bug
when creating with sfdisk, there was a bug where we would not utilize
the maximum amount available if the disk was over 2TB.

This also adds support for GPT disks via sgdisk.

One thing that needs to be added is choice of which to use.
Right now it defaults to gpt without even checking for availability of sgdisk.

bzr-revno: 96