fix searching for eatmydata in target
This fixes a bug that resulted from the behavior of os.path.join.
Instead of searching for 'eatmydata' inside the provided target, we were
only searching inside /.

The fix here is to improve 'which' to support searching inside a
provided target, and then to use which from install_packages.

LP: #1358350
bzr-revno: 148