run 'update-grub' explicitly, re-work setting grub variables.
We need to run update-grub in order to ensure that /boot/grub/grub.cfg
is written.  If that file does not exist, then running dpkg-reconfigure
will not run update-grub.  It is possible that we're doing duplicate
work here in re-running update-grub, but it is minimal.

We've also moved how we're disabling the grub os prober to be permenant.

LP: #1373137
bzr-revno: 177.1.1