tools/launch: fix publishing of non-absolute paths
we were getting the full path with readlink -f, but then
not using it for the target of the link.

also here, use the name provided by the user rather than
the output of readlink -f for the default publish path.  Ie,
if the user did:
  ln -s /tmp/foofile.txt my.txt
we *were* publishing that as foofile.txt rather than my.txt.

bzr-revno: 185