extract: extract without --acl
While this is arguably lossy of acl permissions in the tarball
right now we don't have a better way to do this.

The issue is that if --acl is added, then tar creates default
permissions on the extracted top level directory, which then
are recursively applied to the filesystem.  Random things fail
as a result.

LP: #1382632
bzr-revno: 194