improve code that finds devices and partitions
this fixes 2 bugs (LP: #1401190 and LP: #1263181).
I'm not terribly happy about the name-specific (cciss/c0d0) code that is
being added here.  However, I do want this to work.

Also, we add a "should work" path for determining the device name
of a partition on a given target.  Ie, given:
  /dev/sda 1 => /dev/sda1
  /dev/nbd0p1 => /dev/ndb0p1

The tricky/annoying one is
  /dev/cciss/c0d0 1 -> /dev/cciss/c0d0p1

LP: #1263181, #1401190
bzr-revno: 201.1.1