helpers/common: wipe_dev / wipe_partitions support --full
wipe_partitions and wipe_dev can now take first argument
of a '--full' to wipe the whole thing, not just the front/end.

this can be necessary for PReP, as I've seen grub complain
on install that the PReP partition was not empty.

So, we zero the whole PReP partition.

bzr-revno: 219.1.10