Verify that the HTTP service used by --publish is launched by launch
tools/launch did not check that the backgrounded HTTP service was up
and running.  This led to odd failures in the image since it couldn't
obtain any published data.

The fix requires waiting for the backgrounded job to become active and
attempt to open the desinated port.  launch will wait a total of 2 seconds
to check that the port is active.  If the service is up, we check if we
get the same content from a placed file with a checksum'ed dated input.

This ensures we're getting data from the expected service on the
specified port.

LP: #1489889
bzr-revno: 253.1.1