add 'system-upgrade' command
The intent of this is to be runnable easily as 'curtin system-upgrade'
We can expose this from a config option and run it in curthooks also.

One such reason for this is shown in bug 1481992.
upgrading a system is often times unfortunately more complex than
'apt-get dist-upgrade'.

And the goal would be to make this run the appropriate system upgrade
command for the os in target (yum update for example).

bzr-revno: 269.1.1