xkvm: handle killing qemu-system process if killed
if the xkvm process was killed, other than by a ctrl-c on the terminal,
then we would not kill the qemu-system process.

Now if user runs from console, then run "in foreground".
Otherwise, by default run via '&' and then wait so we can
catch the signal and pass it on.

bzr-revno: 270.2.2