grub-install: if fail to find mounted uefi partition, give more debug output
The code here is probably legacy.  It was added with the original UEFI code
that is now gone.  In fact the 'mp_dev' variable is not used anywhere else.

For the moment we'll leave it here, but add some more debug information in
the event that the check fails.

bzr-revno: 273