subcommands should return value, not sys.exit
I noticed that the reporting code was catching a FAIL operation on
curthooks.  This was because the command was using sys.exit(0), which
would be caught by the 'except' branch of main.

The context of args.reportstack would get the exception and mark the
report event as FAIL.

The changes here are then:
 a.) make all the subcommands return integers rather than calling sys.exit
     Note, some still raise exceptions, which I think is OK for the time
     being.
 b.) make main handle SystemExit specially.
     we WARN and print a traceback as this is just something that we should
     see immediately and fix.

bzr-revno: 273.1.1