maybe small improvements
still have OUTPUT_COLLECT_D as a string some places.
since we're doing 'replace' to fill it in, its just as easy as 
a static constant than a variable at this point

ie, instead of:
  """cd OUTPUT_COLLECT_D"""
the users would have to:
  "cd " + vmtests.OUTPUT_COLLECT_D + ""

i see it 6 in one half a dozen in the other.

bzr-revno: 277.1.3