no longer assume http_proxy means apt_proxy
because the config in hosts apt is more specifically set for an
apt proxy, and the user might have a http_proxy set to something
else, for the moment we will not consider http_proxy in setting
apt_proxy.

its easy enough to set the apt_proxy environment variable
explicitly to http_proxy if you want that.

bzr-revno: 277.2.12