vmtests: ensure tmpdir is cleaned between test class changes
vmtests refactoring exposed an issue with the amount of storage
required to complete a full make vmtest.  At it's peak, it requires
48G of storage space due to the base class __del__ function not
being invoked when switching between testcase classes in a test
file.

This patch fixes this by moving the delete logic into a method
on the TempDir class and adding a call to the new method in to
base class' tearDownClass classmethod.

bzr-revno: 291.2.1