add and use helper for getting boolean value from environment variable
We were doing this:
  if (os.getenv('CURTIN_VMTEST_KEEP_DATA', "false") != "true"):

But that meant that these seemingly false values would be read as true:
   "", "FALSE", "0".

Put that check in one place.  Now, we instead do:
   get_env_var_bool('CURTIN_VMTEST_KEEP_DATA', False):

Which considers any of the values listed above as false, and all others
to be true.  If variable is not present at all, then default is False.

bzr-revno: 272.1.20