improve launcher to provide the check module with args and delim
this sets the variable PY3OR2_ARGS_DELIM and PY3OR2_ARGS when
invoking the MCHECK executable.

As an example, if ./bin/curtin is called with 'mkfs --help',
then we'll invoke:
  PY3OR2_ARGS_DELIM="|" PY3OR2_ARGS="mkfs|--help" \
     python -m curtin.deps.check

This way the called program can inspect the arguments.
This could later allow
  curtin --config=foo.cfg install ...
to only install the dependencies that were necesasary for
the provided config.

bzr-revno: 254.2.13