drop launcher's usage of 'deps.install'
previously, the launcher would invoke
  $python -m deps.install
that would then be in charge of installing dependencies.

Instead of that, now main takes and respects the '--install-deps'
flag.  If you want deps installed automatically, then --install-deps.

Having argparse handle the parsing means that we do not have
to specially handle 'pack' any more.  Ie,
   curtin pack curtin --install-deps install ....
wil do the right thing.  It does not think that it should
install dependencies on this system, but rather pack up that
command to be executed.

bzr-revno: 254.2.35