do not use atexit
i think  maybe the testcase runner is using atexit for teardown
or some other odd ordering, but the dir was not being cleaned up.

probably need to re-think how we want to have something that
runs at the very end that gets access to total  pass/fail
and can act on that.

bzr-revno: 299.1.12