fix curtin.deps.install main
running 'python -m curtin.deps.install'
shows the mismatch between what was in main in curtin/deps/install.py
and in curtin.deps.install_deps.

There isnt any code that currently uses this function, but its likely
we will use in the future, and is a nice way to get dependencies installed.

bzr-revno: 310