implement KEEP_DATA 'fail', simplify test case inheritance
This is probably not the most elegant way to implement this, but seems
to work.  We have our VMBaseClass test override unittest.run and then
look at the result afterwords.  Then, in the test-specific directory
we will have either a 'success' or a 'errors.json' at the end.

in the tearDownClass we can check for success or failure and leave
results around appropriately.

bzr-revno: 299.1.21