initial fix for reading skew_data
I picked these from cloud-init revno 1147 that made similar changes.
One thing I don't yet have here is the retries that went in there.
Essentially you have to retry to put the skew in place.

We probably should get that here too.

There is actually one other set of changes that i'd like to get.
Currently we set the oauth skew, and then use it, but if something
(systemd-timed) updated the clock during these things happening
then our skew could be come wrong.

The fix I want to put in place in both cloud-init and curtin
for that is to store the offset between the uptime() and the server
reported time in seconds since epoch.  Then we can just read the current
uptime and add the stored delta, then convert that to a date and use it.
It wont matter then if something changed the system clock in between.

LP: #1520400
bzr-revno: 314.4.1