vmtests: invoke nosetests3 with --nologcapture
When a install or boot fails, the vmtest harness would log.DEBUG
with the console log output.  nose was capturing this logging
and replaying it on test failure.

This causes very long output to the screen of the user who ran
'make vmtest', making it hard for them to find out what happened.

Since we're now reasonably storing the console logs, its quite
sane to not spit those to stderr so that the output is easier to read.

bzr-revno: 318.1.1