support passing an integer or valid float to human2bytes
We were forcing people to pass 'size' in in a config as a string.
If a 'size' entry had an actual integer, then we'd raise exception.
human2bytes(100) would raise an exception.

Seems to make sense for human2bytes to allow int (100) or float (100.0).
This does check and raise error on 100.4.

bzr-revno: 320.1.1