In http://bazaar.launchpad.net/~curtin-dev/curtin/trunk/revision/202 the devices which curtin will attempt to install on was limited to only devices which are not marked as removable by lsblk. Unfortunately lsblk reports some drives as removable which in reality are the main drives in the server.
In order to still favor non-removable drives but still allow removable drives as a last resort, this commit performs a check after searching for installable and non-removable devices if any were found. If none were found we then search for installable and removable devices which can be installed on. If we still find no devices, an exception is thrown. This exception throws a cleaner error message than the previous error received by users encountering this case, which was a cryptic "list index out of range" error.

bzr-revno: 332.3.1