Further cleanups on syncing images
This clears up some of the unexpected behavior of syncing images.  Much of
the confusion was because sstream-mirror keeps internal information in
.data/ that sstream-query does not read.  The result is that:
  sstream-mirror <remote> <local> <filters>
and then a
  sstream-query <local> <other_filters>

would show results as if <other_filters> had been mirrored even though
the files were not present.

We clean that up.  Now, get_images will ensure that the data you were
after is present.

Also, we no longer attempt mirror on every setUpClass.  That could result
in one Precise test running with one image and the next Precise test
running with another image.  This will only run a mirror once for a given
set of filters during a nosetests run.

bzr-revno: 337.1.1