force use of right python for pylint.
py34-plint when run on c-i was showing:
  E:200,24: Module 'urllib' has no 'parse' member (no-member)
  E:622, 8: Instance of 'VMBaseClass' has no 'assertRegex' member (no-member)

which indicates it was running with python2.
this should force it to run with python3 (the environment python)
in the tox environment.

The reason for the difference from my system I believe is that tox itself
is python2 on trusty and python3 in newer ubuntu.  So the setup runs
in python2 and just invoking 'pylint' then uses the interpreter
on the c-i system is a 

bzr-revno: 338.2.2