vmtest sync: fix pruning of dirs
our pruning code is somewhat simplistic at this point.
We really need to have a better mirror, but for now this fixes an issue.

At some point on our c-i nodes, a mirror included 'di' data, which
creates files in '<mirror>/<release>/<arch>/di'. That directory was 
getting counted as a 'image dir'.

bzr-revno: 340.1.1