vmtests: fix bug in test_dname
test_dname was iterating of a dictionary of items (disk_to_check)
which included the disk's 'name' dname, and then validated that
curtin also emitted a rule to generate a -partX for each partition
created on the disk (if the disk had a dname).

The testcase used the syntax:

disk_to_check = {'main_disk': 1, 'main_disk': 2}

This produces a dictionary with a single key ('main_disk') with a single
value (2).

When interating, the test_dname would check the *last* partition dname
rather than _all_ of the items.

Fix this by using a list of tuples and update test_dname to expect
such a list.

Finally, update all tests that used disk_to_check to use the [tuple()]
syntax to indicate what the testcase expects.

bzr-revno: 351.1.10