add test_simple
This does 2 things:
 a.) provides a 'simple' config that does not do network
     previously all we had was 'basic_network'
     that did do the simple, but also did networking.
 b.) add test_simple to test that works for xenial ppc64

Note the 'cleanup_1436810' is necessary now to work around LP: 1436810.
Note also that the 'xxx_storage' is not used, but could be.
We also should have a storage test that works on power that config
would be a good starting point.

Note also, that at the moment the Trusty test is failing.

bzr-revno: 356.2.3