tox.ini: run with 'python3' rather than python3.4 or python3.5
There is python3 on xenial, but it is python3.5 which differs
from the python3.4 which is on trusty.

This change still allows 'tox' to work on trusty.
Admittedly, we do test python3.5 only on xenial and python3.4 only
on trusty, but we can solve that at a future date.

bzr-revno: 356.1.1