net: don't write udev rules if iface does not have mac_address
For manually configured interfaces, we only require a type: 'physical'
and a name: key in the network config.  In this case, if we have no
mac_address key then we do not need to emit a udev rule.

Add some logging and info output to make apply_net provide useful
output as well as stacktrace upon errors.

bzr-revno: 359.2.3