fix tox running for python2 on trusty
This is similar to commit 350.  We do not run python2 checkers 
against tests/vmtests. 

For some reason this just started failing.  The change is to just
add a py2-flake8 and py3-flake8 environment.  Note, we are not running
py2-flake8 by default right now (its not in the env list)

bzr-revno: 360