curtin: fix tox errors in commands/main.py
Unexpectedly, pylint now blows up with the following errors:

E:208,13: Instance of 'argv' has no 'reportstack' member (but some types could
        not be inferred) (maybe-no-member)
E:209,18: Instance of 'argv' has no 'func' member (but some types could not be
        inferred) (maybe-no-member)

This code has been in tree for a very long time, this only shows up when
we added pylint to the tox run (revno 335).  This should have always failed.

With that said, the fix is simple.  In main(), we define args to be sys.argv
but it's not used until later we re-assign args to be something else from
argparser.  Removing that useless initialization as a argv type resolves
the lint error.

bzr-revno: 365.1.1