block: wipe_volume improvements
- move wipe_volume, find_sysfs_path, from block_meta to block
- improve wipe_volume:
  - do not shell out for 'dd' and ignore errors
  - add 'superblocks' mode that wipes start and end of existing
    partitions in addition to volume.
  - pick 4M as the default size for writes. This is easily modified.
- add block-wipe subcommand for:
  curtin block-wipe --mode=random /dev/vdb /dev/vdc
- sysfs_partition_data: this returns information on partitions on a
  given block device.
- sys_block_path: handle case where devpath is a partition and prepend
  parent dev into path.
  
Note, this does not call rereadpt as sgdisk --zap-all would.
We probably want to add the ability to do that. The reason I dont have it
yet is that in newer ubuntu systemd is watching readwrite opens on the
block devices and will issue rereadpt when those are closed.

bzr-revno: 376