net: introduce 'control' field to network configuration
Allow users to specify how they want ifupdown to treat the
interface.  The default value for control is 'auto' which
results in an auto $iface line in eni.  If control is
'manual', then we emit a comment instead, # control-manual $iface
Any other value of control will emit a an allow-$control value.

These lines control whether ifupdown will ensure the interfaces are
up during boot.  auto and hotplug ifaces will be considered a requirement
for network to be "up".  

The 'manual' setting is useful for things like iscsiroot which
will configure networking in initramfs and when the root filesystem
runs, we do not want ifupdown to bring the interface used to connect
to the iscsi target up or down.

bzr-revno: 380.1.1