simplify apt_source file wirte and read
The curtin.util file functions are much less complex than those of cloud-init.
The imported code had to be adapted.
For the given case that meant dropping all encode/decode features which is ok
since we only read/write text in these places and that is well supported by
curtins utils.

bzr-revno: 389.7.14