multipath testing
Add multipath testing via a test-case option, cls.multipath which will
if set to true, duplicate all disks in a test class.  This has the effect
of looking exactly like multipath setups where we see the same device 
with two paths.

To handle devices with wwn or serial numbers with spaces in them curtin.block
was updated to ensure it supports finding devices via there serial or wwn 
attribute. Specifically for serial numbers with spaces, we currently will
escape the spaces to underscores; this is supported by observing the names
that show up in /dev/disk/by-id/* for those devices.

For wwn, /dev/disk/by-id/wwn-$wwn ; curtin resolves the symlink target to
obtain a kernel devname for use with sysfs.

Finally the multipath test using serial with spaces fails under Xenial due to
LP:1551937 .  This will be resolved in a separate commit.

bzr-revno: 389.5.2