Add tools/curtin-log-print
Include a program for parsing and prettifying reporting output.  Included
is a --blame option to indicate which events take the longest.

Update reporting test-case to assert that a start/finish event pair do not
have the same timestamp; this was an issue with the Reporting class which
only initialized timestamp once at import time.

Add curtin-log-print to make check output.

bzr-revno: 381.2.4