improve net-meta network configuration
when network configuration is not provided 'net-meta' would write
a very simplistic /etc/network/interfaces file for each of the
devices that were brought up in install environment.  This had the following
issues
a.) it did not disable cloud-init networking
b.) it did not account for naming of devices (causing issue if
    devices changed names between install and boot environment).
c.) did not include interfaces.d as ideally it would.

LP: #1592149
bzr-revno: 390.1.1