reporting: default webhook handler level to DEBUG, no filtering
The webhook hanlder in curtin reporting included a default level
of INFO and included a event filter which would not report events
that were of lower importance than what was set.  The result was
for unconfigured WebHook handlers, some of the expected messages
including post_files used by maas to collect the curtin install
log were no longer sent.

This patch sets WebHookHandler level to DEBUG, the same level set for
LogHandler and PrintHandler, further we remove the event filtering.
The events themselves include the level value and the event collector
can apply their own filtering as needed.  Along with the changes we
clean up some comments and include a new unittest which specifically
tests that a FinishReportingEvent with post_files is posted by the
WebHookHandler.

LP: #1590846
bzr-revno: 389.4.1