tools/launch: boot by root=LABEL=cloudimg-rootfs
We were previously booting with either root=/dev/vda or root=/dev/vda1
based on if the root disk was partitioned or not.

That decision logic was sort of broken in that it didn't correctly
query the backing device of a qcow2 disk.

What prompted this change though is that using root= seems not reliable.
On power8 kvm guests on 16.04 in current trunk, root=/dev/vda would not
find a filesystem, where the same boot on 15.10 did.

So for now, boot by LABEL= which is true for all of our vmtests.

bzr-revno: 395.2.1