In block.wipe_volume when running with mode pvremove, do not use '--cache' flag on trusty.
Previously, the cache flag was appended to pvscan an vgscan cmds on
all releases other than precise, as precise is the only ubuntu release
currently supported that has lvm tools that do not have the '--cache' flag as
an option.

However, trusty does not by default use lvmetad, and in cases where
lvmetad is not present or is not running, the pvscan and vgscan tools will
exit with a bad return code (5) and will not scan for volumes.

Not appending the '--cache' flag on trusty is preferrable to accepting a bad
return code from pvscan and vgscan, as there are cases where these tools
failing to refresh the nodes in /dev could halt installation or cause problems
for clear_holders

bzr-revno: 392.2.8