add support for RunInChroot in curtin.util
In my merge committed to trunk in revno 404, we dropped 'RunInChroot'.
That method was being used from curthooks in maas-images project to
make a chrootable target.  In order to support installation of old images
that depend on RunInChroot we need to add the method back.

This does that by just utilizing ChrootableTarget.  It also adds
a couple unit tests to ensure that this keeps working.

LP: #1617375
bzr-revno: 420.1.1