Add workaround for bug 1645680 until resolved in apt or gpg
Most of our gpg related calls run in the host and are not affected.
Only add-apt-repository is in the chroot, due to changes in gpg >=Yakkety
we need to get rid of the dirmngr and gpg-agent daemons before we are able
to umount.

This adds the discussed "workaround" as a fast countermeasure.
The intention is to drop it once (hopefully ever) add-apt-repostory and/or gpg
take care of it.

Since it is just a failsafe addition we want to avoid the potential breakage of
a killall. Theregote it is very sensitive when to kill the daemons.
As discussed it uses --younger-than and on top an optional arg to
ChrootableTarget flag that has to be set.

bzr-revno: 432.1.1