vmtest: add CURTIN_VMTEST_CURTIN_EXE variable.
This allows us to control which 'curtin' program is used.  The program
must support curtin's 'pack' interface, and needs to handle finding itself
appropriately.

Also here, we drop the PYTHONPATH setting in launch, as it is no longer
needed as our bin/curtin wrapper handles that.

The target use case is really for easily running vmtest with an installed
version of curtin, allowing us to easily verify stable release updates.

bzr-revno: 438.1.1