vmtest: overhaul image sync
Changes to support HWE and Centos images have exposed issues in the image sync
process.  Our 'make sync-images' was using a simplestreams filter that did not
pickup any of the required HWE kernels which meant that during vmtest runs
we would trigger new image syncs.  Compounding this issue was the fact that the
filter used was too-wide which picked up things like di-initrd,di-kernel files.

Additionally for an empty IMAGE_DIR repository, there were two bugs; one it
assumed that it could load the vmtest.json/vmtest-centos.json files.  Second
when the repo was empty, we did not return an iterable type.

The following changes have been made to achieve the goals:
  - make sync-images should download everything we need to run a complete
    vmtest run without acquiring any new files from our image server.
  - Handle empty repositories properly
  - Do not attempt to sync with the streamds on each test, but only if specific
    needed files are missing

bzr-revno: 438.3.2