curtin/block/iscsi: require [] around IPv6 addresses
Currently, we do not require [] around IPv6 addresses and attempt to
parse the resulting string correctly (you can get odd misparses as in
the bug). This is clearly error-prone, so let's require that IPv6
addresses must have [] around the address part. We either allow a string
without ':' and '@' in it (and validate it is hostname or IPv4 address
later) or if [] is found we allow a string without '@' in it and
validate it is an IPv6 address.
      
This ends up allowing for ':' in target names which otherwise gets
mis-parsed as in LP #1679222.

Fixes: LP: #1679222.

bzr-revno: 481.1.2