add recursive flag to do_umount, add --umount to block-meta
This adds a 'recursive' argument to util.do_umount, and utilizes
it from install.py.  The basic idea of do_umount with recursive
is to walk through /proc/mounts in reverse and unmount anything
mounted under the provided mountpoint (or the mountpoint itself).

Also added here is use of that function from block-meta command.
We add a flag --umount to unmount filesystems at its end.

bzr-revno: 483.2.1