In bcache test, run block-meta once inside the early_commands.
This runs 'curtin block-meta' from the curtin install's early_commands.
So essentially, the early command partition the disks, put a filesystem
on them, and then the install continues and those are all wiped away.

It functions as a somewhat simple way to get "dirty" disks in place, and
in fact actually catches bug 1676991 (bcache race condition) if we
leave the '-smp 2' modification in launch.

bzr-revno: 483.2.2