clear-holders: mdadm use /proc/mdstat to wait until an array has been stopped
shutdown_mdadm needs to wait until the md device has stopped.  Using files
in sysfs is unreliable due to a kernel bug (LP:1682456) so instead use
device presence in /proc/mdstat.  

Add the use of --manage flag to force mdadm to interrupt actions on the device
(like a resync).

bzr-revno: 437.2.11