tests: clean up usage of collect files
This adds a helper 'collect_path' to the vmtests base class, and then
uses it wherever appropriate.  It also cleans up all code that
used read collect files to  use load_collect_file.

Lastly, it updates output_files_exist and output_files_dont_exist
to show which files do or do not exist.  Previously all you'd see
in the log was an assertion that something was or was not true.
Now, you'll see a list of files that do or do not exist, and not
just the first.

bzr-revno: 489.4.1