vmtest: let the socket level select a port rather than us picking one.
Previously we had some code in tools/launch to pick a random port
from a range of ports.  That worked for the most part, but we recently
had a failure in vmtest where we failed to select a port in 5 tries.

Instead of selecting a port at random and then trying to bind to it, we
can ask the socket layer for port 0, and it automatically assigns one.
We then modify 'webserv' to output the port it binded and 'launch' to
read it.

bzr-revno: 504.1.1