tools/launch: automatically pass on proxy settings to curtin
This makes tools/launch automatically pass on proxy settings
  http_proxy, https_proxy, no_proxy
from its environment into curtin via curtins' newly added 'proxy' config.

It also puts the discovered IP address into 'no_proxy'.  This means
that by default, curtin will get the host environment's proxy settings
but will not attempt to use the proxy for the host's IP address. This is
done because it is likely that the configured proxy will not proxy
traffic back to this host.

As an implementation detail, it adds 'curtin_extra' arguments and
integrates the previously added '--install-deps' flag into that.

bzr-revno: 508.1.1