tools/jenkins-runner: improve tgtd cleanup logic
The previous tgtd cleanup logic was hard to follow at best.  It was buggy
in that it would always kill the TGT_PID process even if it didn't start it.
both cleanup lines ended with:
  <condition> || [ -n "${TGT_PID}" ] && kill -9 ${TGT_PID}
which will always run the 'kill' if TGT_PID is non-empty.

The new code will kill only tgtd processes that it started, and will
keep the TGT_LOG_D if the "keep rules" were indicate 'logs'.
It will not kill the TGT_PID process that it started if 'tgtd' is in
the "keep rules".

Also clean up the output.

bzr-revno: 508.2.1