separate out the wait-for logic
this allows us to re-try on empty port.
i have no proof, but suspect that we were just looking at netstat
before tgtd had a chance to come up and listen to the port.

this way, if *nothing* is listening, we sleep and look again.
if something else is listening, kill that try.

bzr-revno: 508.2.10