tests: in centos_defaults use write_files for grub serial.
This changes centos defaults to pass a file into the target
with write_files and then use that program in late_commands to
modify grub1 and grub2 config files so that we get console
output to the serial console.

Instead of running grub2-mkconfig it just updates both
the defaults file and the rendered /boot/grub2/grub.cfg.

bzr-revno: 511.2.1