centos_network_curthooks -> centos_apply_network_config.
Rename centos_network_curthooks and its upper-cased feature
CENTOS_NETWORK_CURTHOOKS to centos_apply_network_config and
CENTOS_APPLY_NETWORK_CONFIG respectively.

Also drop the handling of 'cloudconfig' entries from that function.
Handling of cloudconfig is up to the curthooks, and that is being
added to maas-images built-in curthooks.

The name centos_apply_network_config is tested
We will leave that in curtin.cmd.curthooks and will add unit
test to that that ensures that entry point is never lost.

Also add unit test to ensure 'write_files' is available from futil.


bzr-revno: 511.3.1