vmtest: get info about collected symlinks and then delete them.
Some of the 'collect' scripts collect  files with 'cp -a' or other
means that would collect a symbolic link.   Since none of our
scripts yet handle properly resolving absolute symlinks we just
remove all symlinks and write a file 'symlinks.txt' with info about them.

At a later date we might need to also clean out sockets or other content.

The primary motivation for this is to
a.) make sure tests don't inadvertantly read the target of the symlink
    and thus "verify" files on the host
b.) jenkins can produce a zip file of the artifacts, and we suspect that
    it is bothered by the symbolic links.

bzr-revno: 519.2.1