block: enable control over exclusive_open use when wiping volumes
The exclusive_open check on a device is very useful to determine if
curtin is removing all metadata from a device w.r.t storage configuraiton.
However, in some cases during creating of new storage configations as
we attempt to add additional partitions to a device which has been
added to a md or lvm or bcache layer, the exclusive open is blocked while
we attempt to zero some data on the underlying device.

This patch exposes a flag to the zero_file_at_offset method to allow callers
to disable the exclusive_open.  This lets curtin wipe data at specific offsets
on a volume without requiring exclusive ownership of the device.

bzr-revno: 537.1.1